% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_shots.R
\name{analyse_shots}
\alias{analyse_shots}
\title{Bar chart and tibble of shot type in the 2001-2002 and 2021-2022 seasons filtered by NBA team}
\usage{
analyse_shots(data, team_filter = NULL, metric = c("shots", "points"))
}
\value{
\code{ggplot2} a summary tible and side by side bar charts
}
\description{
Creates two bar charts and tibbles that show either the total number of points from 2 and 3-point attempts
or the total number of shots, filtered by team and split by season from NBA data
}
\examples{
analyse_shots(data = clean_data, team_filter = "BOS", metric = "points")$plot
analyse_shots(data = clean_data, team_filter = "ATL", metric = "shots")$summary

}
